refactor(wiki): split pg_store_wiki.py under the 300-line rule
pg_store_wiki.py was 890 lines, ~3x the CLAUDE.md 300-line-per-file
hard rule. Split along wiki sub-entity concern boundaries (Extract
Module / Move Function, Fowler):

- pg_store_wiki_common.py (32 lines) — body_hash, _returning_id helpers
- pg_store_wiki_pages.py (143 lines) — wiki.pages CRUD
- pg_store_wiki_links.py (80 lines) — wiki.links
- pg_store_wiki_claims.py (205 lines) — wiki.claim_events
- pg_store_wiki_thermo.py (126 lines) — bulk thermodynamic/staleness updates
- pg_store_wiki_concepts.py (108 lines) — wiki.concepts
- pg_store_wiki_drafts.py (185 lines) — wiki.drafts
- pg_store_wiki_notes.py (91 lines) — wiki.citations / wiki.memos / diagnostics
- pg_store_wiki.py (108 lines) — thin re-export facade, keeps all 9
  existing importers (wiki_refine/curate/synthesize/emerge/resolve/
  extract/migrate/consolidate/compile) working unchanged

No logic changed — every function moved verbatim (SQL, docstrings,
type hints byte-for-byte). Public API (34 real names) verified
identical via dir() comparison before/after; only incidental re-export
names (Any/Connection/dict_row/hashlib/json, never imported by any
caller) differ.

Before: 890 lines / 34 public names.
After: 9 files, all <=300 lines / same 34 public names.
Tests: 4181 passed / 1 failed before change too (pre-existing
order-dependent flaky integration test, confirmed passing in isolation
both before and after this refactor) / 0 tests modified.
ruff check: clean on all touched/created files.
